requirements.txt
asgiref==3.5.0
boto3==1.22.9
botocore==1.25.9
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
defusedxml==0.7.1
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
django-storages==1.12.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
gunicorn==20.1.0
idna==3.3
jmespath==1.0.0
numpy==1.22.3
oauthlib==3.2.0
Pillow==9.1.0
pycparser==2.21
PyJWT==1.7.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.1.0
